Moved tests from integration to unit (I think appropriately) so they are picked up by codecov.
